requirements.txt
smartapi
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.6.0
flask-pymongo==2.3.0
requests==2.32.3
APScheduler==3.10.4
Werkzeug==3.0.1
python-dotenv==1.0.0
flask-limiter==3.5.0
twilio==8.10.0
smartapi-python==1.4.0
pyotp==2.9.0
websocket-client==0.59.0
logzero==1.7.0
yfinance==0.2.40
pycryptodome==3.20.0  